Split cancel run command
